requirements.txt
pandas
matplotlib
torch
seaborn
mplsoccer
python-dotenv